+++ src/OpenTelemetry.Instrumentation.Owin/CHANGELOG.md
@@ -6,6 +6,8 @@
   ([#1344](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1344))
 * Removes `AddOwinInstrumentation` method with default configure parameter.
   ([#929](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/929))
+* Adds HTTP server metrics via `AddOwinInstrumentation` extension method on `MeterProviderBuilder`
+  ([#1335](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1335))
 
 ## 1.0.0-rc.3